Relocate cpp compile harness to module
This file is not supposed to be a separate integration test crate. It is
only intended to be used as a module. Otherwise it is compiled twice.
